Revert ":sparkles: Todas las rutas completadas"
