+++ server/src/DB_Connection.js
@@ -1,7 +1,7 @@
 const { Sequelize } = require("sequelize");
 const { DB_USER, DB_PASSWORD, DB_HOST } = process.env;
 const UserModel = require("./models/User");
-const FavoriteModel = require("./models/Favorites");
+const FavoriteModel = require("./models/Favorite");
 
 const sequelize = new Sequelize(
   `postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}/rickandmorty`,
