+++ server/src/models/Favorite.js
