--- server/src/routes/charactersRouter.js
@@ -1,8 +1,0 @@
-const { Router } = require("express");
-const getCharacterHandler = require("../handlers/charactersHandler");
-
-const charactersRouter = Router();
-
-charactersRouter.get("/:id", getCharacterHandler);
-
-module.exports = charactersRouter;
