--- server/src/routes/favoritesRouter.js
@@ -1,12 +1,0 @@
-const { Router } = require("express");
-const {
-  postFavoriteHandler,
-  deleteFavoriteHandler,
-} = require("../handlers/favoritesHandler");
-
-const favoritesRouter = Router();
-
-favoritesRouter.post("/", postFavoriteHandler);
-favoritesRouter.delete("/:id", deleteFavoriteHandler);
-
-module.exports = favoritesRouter;
