+++ server/src/routes/userRouter.js
@@ -1,0 +1,8 @@
+const { Router } = require("express");
+const { postUserHandler } = require("../handlers/UserHandler");
+
+const userRouter = Router();
+
+userRouter.post("/register", postUserHandler);
+
+module.exports = userRouter;
